requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
colorlog==6.6.0
croniter==1.3.5
forbiddenfruit==0.1.4
frozenlist==1.3.1
hikari==2.0.0.dev110
hikari-crescent==0.3.3
idna==3.4
multidict==6.0.2
mypy-extensions==0.4.3
PSNAWP==1.0.1
pycooldown==0.1.0b8
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
sigparse==1.3.1
six==1.16.0
types-croniter==1.3.2
urllib3==1.26.12
yarl==1.8.1
